Update header and button styles
Modified the CSS styling of the website header and buttons. The header height in the startpage/index.html has been set to 100vh to ensure it takes up the full viewport height. Also, the CSS for the buttons has been updated to use transparent background colors, to improve text legibility.